move badge component to separate file
