+++ packages/ui/tsconfig.json
@@ -3,6 +3,7 @@
   "include": ["src/**/*", "tests/**/*", "vitest.config.ts"],
   "compilerOptions": {
     "noEmit": true,
+    "jsx": "react-jsx",
     "baseUrl": "."
   },
   "references": [
